added css modules plugin
