+++ package.json
@@ -26,10 +26,11 @@
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.26.0",
-    "node-sass": "^6.0.1",
     "prettier": "^2.4.1",
     "react-scripts": "4.0.3",
-    "typescript": "^4.4.3"
+    "sass": "^1.52.3",
+    "typescript": "^4.4.3",
+    "typescript-plugin-css-modules": "^3.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
