+++ src/components/Router/Router.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Route, Switch, BrowserRouter } from 'react-router-dom';
 
-import { HomePage, RoomPage, NotFoundPage } from '../pages';
+import { LandingPage, RoomPage, NotFoundPage } from '../pages';
 
 interface Router {
     routes: { [key: string]: Route };
@@ -11,7 +11,7 @@ const Router = (): React.ReactElement => {
     return (
         <BrowserRouter>
             <Switch>
-                <Route component={HomePage} exact path="/" />
+                <Route component={LandingPage} exact path="/" />
                 <Route component={RoomPage} path="/room/:roomId" />
                 <Route component={NotFoundPage} />
             </Switch>
