--- src/components/pages/HomePage/HomePage.scss
@@ -1,20 +1,0 @@
-.home-page {
-    display: flex;
-    height: 100%;
-
-    h1 {
-        font-size: 3em;
-    }
-    
-    section {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        justify-content: center;
-        width: 50vw;
-        background-color: #ddd;
-        &:hover {
-            background-color: #eee;
-        }
-    }
-}
