+++ src/components/pages/LandingPage/LandingPage.module.scss
@@ -1,0 +1,37 @@
+@import '../../../styles/variables.scss';
+ 
+
+.landingPage {
+    height: 100%;
+    width: 100%; 
+    background-color: $color-background;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+
+    .navbar {
+        width: 100%;
+        max-width: $page-max-width;
+        flex-basis: 50px;
+        border: 1px solid black;
+    }
+    
+
+    .main {
+        width: 100%;
+        max-width: $page-max-width;
+        flex-grow: 1;
+
+        .title1 {
+            font-size: 65px;
+            color: $color-primary-contrast;
+        }
+    
+        .title2 {
+            font-size: 22px;
+            color: $color-primary;
+    
+        }
+    }
+    
+}
