+++ src/styles/variables.scss
@@ -1,0 +1,14 @@
+
+$color-primary: #285b9e;
+$color-primary-light: #3375cc;
+$color-primary-contrast: #ffffff; // text written over primary
+
+$color-secondary: #333333;
+$color-seconrady-light: #4d4d4d;
+$color-secondary-contrast: #ffffff; // text written over secondary
+
+$color-background: #e6ecf0;
+$color-background-light: #ebeff1;
+
+
+$page-max-width: 1600px
